add the diagram
convert to Notifier message.
Tobias's feedback.

TBD

discrepancy with RFC 7383 (IKEv2 fragmentation)

If the responder receives a retransmitted fragment of a request when
it has already processed that request and has sent back a response,
that event MUST only trigger a retransmission of the response message
(fragmented or not) if the Fragment Number field in the received
fragment is set to 1; otherwise, it MUST be ignored.